+++ gateway/src/test/java/ru/practicum/shareit/ShareItTests.java
@@ -1,0 +1,10 @@
+package ru.practicum.shareit;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+class ShareItTests {
+    @Test
+    void contextLoads() {
+    }
+}
