+++ server/src/main/java/ru/practicum/shareit/item/entity/model/CommentDto.java
@@ -3,7 +3,6 @@ package ru.practicum.shareit.item.entity.model;
 import lombok.Builder;
 import lombok.Data;
 
-import javax.validation.constraints.NotEmpty;
 import java.io.Serializable;
 import java.sql.Timestamp;
 
@@ -11,7 +10,6 @@ import java.sql.Timestamp;
 @Builder
 public class CommentDto implements Serializable {
     private long id;
-    @NotEmpty(message = "Комментарий не может быть пустым")
     private String text;
     private String authorName;
     private Timestamp created;
