+++ server/src/main/java/ru/practicum/shareit/item/entity/model/ItemDto.java
@@ -5,8 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 import java.util.List;
 
@@ -17,11 +15,8 @@ import java.util.List;
 public class ItemDto implements Serializable {
     private long id;
     private long owner;
-    @NotBlank(message = "Наименование не может быть пустым.")
     private String name;
-    @NotBlank(message = "Описание не может быть пустым.")
     private String description;
-    @NotNull(message = "Статус не может быть пустой.")
     private Boolean available;
     private ItemDtoBookingInfo lastBooking;
     private ItemDtoBookingInfo nextBooking;
