+++ server/src/main/java/ru/practicum/shareit/request/entity/model/ItemRequestDto.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Data;
 import ru.practicum.shareit.item.entity.model.ItemDtoRequestInfo;
 
-import javax.validation.constraints.NotBlank;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -12,7 +11,6 @@ import java.util.List;
 @Builder
 public class ItemRequestDto {
     private Long id;
-    @NotBlank(message = "Описание не может быть пустым.")
     private String description;
     private LocalDateTime created;
     private List<ItemDtoRequestInfo> items;
