+++ server/src/main/java/ru/practicum/shareit/user/entity/model/UserDto.java
@@ -5,8 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
 import java.io.Serializable;
 
 @Data
@@ -16,7 +14,5 @@ import java.io.Serializable;
 public class UserDto implements Serializable {
     private Long id;
     private String name;
-    @NotBlank(message = "Электронная почта не может быть пустой.")
-    @Email(regexp = "^[a-zA-Z0-9_!#$%&'*+/=?`{|}~^.-]+@[a-zA-Z0-9.-]+$", message = "Электронная почта должна соответствовать формату RFC 5322.")
     private String email;
 }
