+++ server/src/test/java/ru/practicum/shareit/booking/entity/BookingTest.java
@@ -13,16 +13,9 @@ import ru.practicum.shareit.item.entity.Item;
 import ru.practicum.shareit.item.entity.model.ItemDtoBookingResponse;
 import ru.practicum.shareit.user.entity.User;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
 import java.sql.Timestamp;
-import java.time.LocalDateTime;
-import java.util.Set;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @JsonTest
 class BookingTest {
@@ -117,54 +110,5 @@ class BookingTest {
         assertThat(result).extractingJsonPathStringValue("$.start").isEqualTo("2022-10-10T10:10:00");
         assertThat(result).extractingJsonPathStringValue("$.end").isEqualTo("2023-10-10T10:10:00");
         assertThat(result).extractingJsonPathStringValue("$.status").isEqualTo(BookingStatus.WAITING.toString());
-    }
-
-    @Test
-    @DisplayName("Создаем проверяемый объект без ошибок валидации.")
-    void whenCreateValidBookingDtoThenCreated() {
-
-        final BookingDtoResponse bookingDtoResponse = BookingDtoResponse.builder()
-                .itemId(1)
-                .start(LocalDateTime.of(2024, 1, 1, 1, 1))
-                .end(LocalDateTime.of(2025, 1, 1, 1, 1))
-                .build();
-
-        //test
-        assertEquals(0, testingValidator(bookingDtoResponse).size(), "Список должен быть пустой.");
-    }
-
-    @Test
-    @DisplayName("Создаем проверяемый объект с ошибками валидации.")
-    void whenCreateValidBookingDtoThenNotCreated() {
-
-        BookingDtoResponse bookingDtoResponse = BookingDtoResponse.builder()
-                .itemId(1)
-                .start(LocalDateTime.of(2000, 1, 1, 1, 1))
-                .end(LocalDateTime.of(2001, 1, 1, 1, 1))
-                .build();
-
-        Set<ConstraintViolation<BookingDtoResponse>> errors2 = testingValidator(bookingDtoResponse);
-        errors2.stream().map(ConstraintViolation::getMessage).forEach(System.out::println);
-
-        //test
-        assertEquals(2, errors2.size(), "Размер списка должен быть равен 2.");
-
-        bookingDtoResponse = BookingDtoResponse.builder()
-                .itemId(1)
-                .start(LocalDateTime.of(2024, 1, 1, 1, 1))
-                .end(LocalDateTime.of(2023, 1, 1, 1, 1))
-                .build();
-
-        Set<ConstraintViolation<BookingDtoResponse>> errors1 = testingValidator(bookingDtoResponse);
-        errors1.stream().map(ConstraintViolation::getMessage).forEach(System.out::println);
-
-        assertEquals(1, errors1.size(), "Размер списка должен быть равен 1.");
-
-    }
-
-    private Set<ConstraintViolation<BookingDtoResponse>> testingValidator(BookingDtoResponse bookingDtoResponse) {
-        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-        Validator validator = factory.getValidator();
-        return validator.validate(bookingDtoResponse);
     }
 }
