+++ server/src/test/java/ru/practicum/shareit/item/entity/ItemTest.java
@@ -12,14 +12,7 @@ import ru.practicum.shareit.item.entity.model.ItemDtoBookingResponse;
 import ru.practicum.shareit.item.entity.model.ItemDtoRequestInfo;
 import ru.practicum.shareit.user.entity.User;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
-import java.util.Set;
-
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @JsonTest
 class ItemTest {
@@ -140,46 +133,5 @@ class ItemTest {
         //test
         assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(1);
         assertThat(result).extractingJsonPathStringValue("$.name").isEqualTo("test");
-    }
-
-    @Test
-    @DisplayName("Создаем проверяемый объект без ошибок валидации.")
-    void whenCreateValidItemDtoThenCreated() {
-
-        final ItemDto itemDtoTest = ItemDto.builder()
-                .id(0L)
-                .name("TestName1")
-                .description("testDescription1")
-                .owner(1L)
-                .available(true)
-                .build();
-
-        //test
-        assertEquals(0, testingValidator(itemDtoTest).size(), "Список должен быть пустой.");
-    }
-
-    @Test
-    @DisplayName("Создаем проверяемый объект с ошибками валидации.")
-    void whenCreateValidItemDtoThenNotCreated() {
-
-        final ItemDto itemDtoTest = ItemDto.builder()
-                .id(0L)
-                .name("")
-                .description("")
-                .owner(1L)
-                .available(null)
-                .build();
-
-        Set<ConstraintViolation<ItemDto>> errors = testingValidator(itemDtoTest);
-        errors.stream().map(ConstraintViolation::getMessage).forEach(System.out::println);
-
-        //test
-        assertEquals(3, errors.size(), "Размер списка должен быть равен 3.");
-    }
-
-    private Set<ConstraintViolation<ItemDto>> testingValidator(ItemDto itemDto) {
-        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-        Validator validator = factory.getValidator();
-        return validator.validate(itemDto);
     }
 }
